requirements.txt
asgiref==3.5.2
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.6.0
Django==4.0.5
django-admin-inline-paginator==0.3.0
django-jalali-date==1.0.1
django-weasyprint==2.1.0
fonttools==4.33.3
html5lib==1.1
jdatetime==4.1.0
Pillow==9.1.1
pycparser==2.21
pydyf==0.2.0
pyphen==0.12.0
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
weasyprint==55.0
webencodings==0.5.1
zopfli==0.2.1
